+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:cdparty_flutter/screens/home_screen.dart';
+import 'package:cdparty_flutter/screens/chat_screen.dart';
 
 void main() => runApp(MyApp());
 
@@ -13,7 +13,7 @@ class MyApp extends StatelessWidget {
         primaryColor: Colors.red,
         accentColor: Colors.orange,
       ),
-      home: HomeScreen(),
+      home: ChatScreen(),
     );
   }
 }
