+++ lib/model/chat_message.dart
@@ -1,0 +1,8 @@
+enum ChatMessageType { incoming, outgoing }
+
+class ChatMessage {
+  final String message;
+  final ChatMessageType type;
+
+  ChatMessage(this.message, this.type);
+}
